Numerical and Symbolic Divergence

Computes the divergence of functions, expressions and characters.

divergence(
  f,
  var,
  coordinates = "cartesian",
  accuracy = 2,
  stepsize = NULL,
  drop = TRUE,
  ...
)

f %divergence% var

Arguments

f
function, expression or character array.

var
character vector, giving the variable names with respect to which derivatives will be computed. If a named vector is provided, derivatives will be computed at that point.

coordinates
coordinate system to use. One of: cartesian, polar, spherical, cylindrical, parabolic, parabolic-cylindrical or a character vector of scale factors for each varibale.

accuracy
accuracy degree for numerical derivatives.

stepsize
finite differences stepsize for numerical derivatives. Auto-optimized by default.

drop
drop dimensions when...

...
additinal arguments passed to f, when f is a function.

Value

divergence array.

Functions

divergence: arbitrary coordinate system

%divergence%: cartesian coordinates

Examples

# divergence of a vector field
f <- c('x^2','y^3','z^4')
divergence(f, var = c('x','y','z'))

#> [1] "2 * x + 3 * y^2 + 4 * z^3"

f %divergence% c('x','y','z')

#> [1] "2 * x + 3 * y^2 + 4 * z^3"

# numerical divergence of a vector field
f <- function(x,y,z) c(x^2, y^3, z^4)
divergence(f, var = c('x'=1,'y'=1,'z'=1))

#> [1] 9

f %divergence% c('x'=1,'y'=1,'z'=1)

#> [1] 9

# divergence in polar coordinates
f <- c('sqrt(r)/10','sqrt(r)')
divergence(f, var = c('r','phi'), coordinates = 'polar')

#> [1] "(0.5 * r^-0.5/10 * r + (sqrt(r)/10)) / (1*r)"