Taylor Series

Computes the Taylor series for functions, expressions or characters.

taylor(f, var = "x", order = 1, accuracy = 2, stepsize = NULL)

Arguments

f
function, expression or character

var
character. The variables of f.

order
the order of the Taylor approximation.

accuracy
accuracy degree for numerical derivatives.

stepsize
finite differences stepsize for numerical derivatives. Auto-optimized by default.

Value

list with components

f

the Taylor series.

order

the approximation order.

terms

data.frame containing the variables, coefficients and degrees of each term in the Taylor series.

Examples

# univariate taylor series 
taylor('exp(x)', var = 'x', order = 3)

#> $f
#> [1] "(1) * 1 + (1) * x^1 + (0.5) * x^2 + (0.166666666666667) * x^3"
#> 
#> $order
#> [1] 3
#> 
#> $terms
#>   var      coef degree
#> 0   1 1.0000000      0
#> 1 x^1 1.0000000      1
#> 2 x^2 0.5000000      2
#> 3 x^3 0.1666667      3
#> 

# univariate taylor series of arbitrary functions
taylor(function(x) exp(x), var = 'x', order = 3)

#> $f
#> [1] "(1) * 1 + (1.00000000248338) * x^1 + (0.499999999915303) * x^2 + (0.166769224538825) * x^3"
#> 
#> $order
#> [1] 3
#> 
#> $terms
#>   var      coef degree
#> 0   1 1.0000000      0
#> 1 x^1 1.0000000      1
#> 2 x^2 0.5000000      2
#> 3 x^3 0.1667692      3
#> 

# multivariate taylor series 
taylor('sin(x*y)', var = c('x','y'), order = 6)

#> $f
#> [1] "(1) * x^1*y^1 + (-0.166666666666667) * x^3*y^3"
#> 
#> $order
#> [1] 6
#> 
#> $terms
#>         var       coef degree
#> 0,0       1  0.0000000      0
#> 0,1     y^1  0.0000000      1
#> 1,0     x^1  0.0000000      1
#> 0,2     y^2  0.0000000      2
#> 2,0     x^2  0.0000000      2
#> 1,1 x^1*y^1  1.0000000      2
#> 0,3     y^3  0.0000000      3
#> 3,0     x^3  0.0000000      3
#> 1,2 x^1*y^2  0.0000000      3
#> 2,1 x^2*y^1  0.0000000      3
#> 0,4     y^4  0.0000000      4
#> 4,0     x^4  0.0000000      4
#> 1,3 x^1*y^3  0.0000000      4
#> 3,1 x^3*y^1  0.0000000      4
#> 2,2 x^2*y^2  0.0000000      4
#> 0,5     y^5  0.0000000      5
#> 5,0     x^5  0.0000000      5
#> 1,4 x^1*y^4  0.0000000      5
#> 4,1 x^4*y^1  0.0000000      5
#> 2,3 x^2*y^3  0.0000000      5
#> 3,2 x^3*y^2  0.0000000      5
#> 0,6     y^6  0.0000000      6
#> 6,0     x^6  0.0000000      6
#> 1,5 x^1*y^5  0.0000000      6
#> 5,1 x^5*y^1  0.0000000      6
#> 2,4 x^2*y^4  0.0000000      6
#> 4,2 x^4*y^2  0.0000000      6
#> 3,3 x^3*y^3 -0.1666667      6
#> 

# multivariate taylor series of arbitrary functions
taylor(function(x,y) sin(x*y), var = c('x','y'), order = 6)

#> $f
#> [1] "(0.999999997516473) * x^1*y^1 + (4.44089209850062e-15) * x^1*y^2 + (4.44089209850062e-15) * x^2*y^1 + (-0.00184136771422416) * x^1*y^3 + (-0.00184136771422416) * x^3*y^1 + (5.91645678915759e-30) * x^2*y^2 + (-3.01980662698043e-14) * x^1*y^4 + (-3.01980662698043e-14) * x^4*y^1 + (-5.9211894646675e-15) * x^2*y^3 + (-3.46944695195361e-18) * x^3*y^2 + (2.04897128321946e-05) * x^1*y^5 + (2.04897128321946e-05) * x^5*y^1 + (-1.22273440309257e-29) * x^2*y^4 + (-8.67746995743112e-30) * x^4*y^2 + (-0.166154533034737) * x^3*y^3"
#> 
#> $order
#> [1] 6
#> 
#> $terms
#>         var          coef degree
#> 0,0       1  0.000000e+00      0
#> 0,1     y^1  0.000000e+00      1
#> 1,0     x^1  0.000000e+00      1
#> 0,2     y^2  0.000000e+00      2
#> 2,0     x^2  0.000000e+00      2
#> 1,1 x^1*y^1  1.000000e+00      2
#> 0,3     y^3  0.000000e+00      3
#> 3,0     x^3  0.000000e+00      3
#> 1,2 x^1*y^2  4.440892e-15      3
#> 2,1 x^2*y^1  4.440892e-15      3
#> 0,4     y^4  0.000000e+00      4
#> 4,0     x^4  0.000000e+00      4
#> 1,3 x^1*y^3 -1.841368e-03      4
#> 3,1 x^3*y^1 -1.841368e-03      4
#> 2,2 x^2*y^2  5.916457e-30      4
#> 0,5     y^5  0.000000e+00      5
#> 5,0     x^5  0.000000e+00      5
#> 1,4 x^1*y^4 -3.019807e-14      5
#> 4,1 x^4*y^1 -3.019807e-14      5
#> 2,3 x^2*y^3 -5.921189e-15      5
#> 3,2 x^3*y^2 -3.469447e-18      5
#> 0,6     y^6  0.000000e+00      6
#> 6,0     x^6  0.000000e+00      6
#> 1,5 x^1*y^5  2.048971e-05      6
#> 5,1 x^5*y^1  2.048971e-05      6
#> 2,4 x^2*y^4 -1.222734e-29      6
#> 4,2 x^4*y^2 -8.677470e-30      6
#> 3,3 x^3*y^3 -1.661545e-01      6
#>