Numerical and Symbolic Hessian

Computes the hessian matrix of functions, expressions and characters.

hessian(f, var, accuracy = 2, stepsize = NULL, coordinates = "cartesian")

f %hessian% var

Arguments

f
function, expression or character.

var
character vector, giving the variable names with respect to which derivatives will be computed. If a named vector is provided, derivatives will be computed at that point.

accuracy
accuracy degree for numerical derivatives.

stepsize
finite differences stepsize for numerical derivatives. Auto-optimized by default.

coordinates
coordinate system to use. One of: cartesian, polar, spherical, cylindrical, parabolic, parabolic-cylindrical or a character vector of scale factors for each varibale.

Value

hessian matrix.

Functions

hessian: arbitrary coordinate system

%hessian%: cartesian coordinates

Examples

# hessian with respect to x
hessian(f = "sin(x)", var = "x")

#> [1] "-sin(x)"

"sin(x)" %hessian% "x"

#> [1] "-sin(x)"

# hessian with respect to x and evaluate in x = 0
hessian(f = "sin(x)", var = c("x" = 0))

#> [1] 0

"sin(x)" %hessian% c(x=0)

#> [1] 0

# hessian with respect to (x,y)
hessian(f = "y*sin(x)", var = c("x","y"))

#>      [,1]            [,2]    
#> [1,] "-(y * sin(x))" "cos(x)"
#> [2,] "cos(x)"        "0"     

"y*sin(x)" %hessian% c("x","y")

#>      [,1]            [,2]    
#> [1,] "-(y * sin(x))" "cos(x)"
#> [2,] "cos(x)"        "0"     

# hessian in spherical coordinates
hessian('r*theta*phi', var = c('r','theta','phi'), coordinates = 'spherical')

#>      [,1]                                                                              
#> [1,] "0"                                                                               
#> [2,] "1/1 * (1/r * phi - 1/r^2 * (r * phi))"                                           
#> [3,] "1/1 * (1/(r * sin(theta)) * theta - sin(theta)/(r * sin(theta))^2 * (r * theta))"
#>      [,2]                                                                              
#> [1,] "1/r * (1/1 * phi)"                                                               
#> [2,] "0"                                                                               
#> [3,] "1/r * (1/(r * sin(theta)) * r - r * cos(theta)/(r * sin(theta))^2 * (r * theta))"
#>      [,3]                              
#> [1,] "1/(r*sin(theta)) * (1/1 * theta)"
#> [2,] "1/(r*sin(theta)) * (1/r * r)"    
#> [3,] "0"                               

# numerical hessian in spherical coordinates
f <- function(r, theta, phi) r*theta*phi
hessian(f, var = c('r'=1, 'theta'=pi/4, 'phi'=pi/4), coordinates = 'spherical')

#>              [,1]         [,2]         [,3]
#> [1,] 1.023182e-12 7.853982e-01 1.110721e+00
#> [2,] 4.545253e-13 2.273737e-13 1.414214e+00
#> [3,] 6.428191e-13 3.034928e-01 4.547474e-13