Numerical Integration

Integrates multidimensional functions, expressions, and characters in arbitrary orthogonal coordinate systems.

integral(
  f,
  bounds,
  relTol = 0.01,
  absTol = 0.001,
  coordinates = "cartesian",
  method = "mc",
  verbose = TRUE,
  ...
)

Arguments

f
function, expression or character.

bounds
list of integration bounds.

relTol
relative accuracy requested.

absTol
absolute accuracy requested.

coordinates
coordinate system to use. One of: cartesian, polar, spherical, cylindrical, parabolic, parabolic-cylindrical or a character vector of scale factors for each varibale.

method
the method to use. Should be one of "mc", "hcubature", "pcubature", "cuhre", "divonne", "suave" or "vegas". Naive Monte Carlo integration by default. The additional methods require the cubature package to be installed (efficient integration in C).

verbose
logical. Print on progress?

...
additional arguments passed to the cubintegrate function, when method "hcubature", "pcubature", "cuhre", "divonne", "suave" or "vegas" is used.

Value

list with components

value

the final estimate of the integral.

abs.error

estimate of the modulus of the absolute error.

cuba

cubature output when method "hcubature", "pcubature", "cuhre", "divonne", "suave" or "vegas" is used.

Examples

# integrate character
integral('sin(x)', bounds = list(x = c(0,2*pi)))

#> n.iter = 1e+05 
#>    value = 0.016; abs.error = 1.4e-02 
#> n.iter = 3e+05 
#>    value = 0.00183; abs.error = 8.1e-03 
#> n.iter = 7e+05 
#>    value = -0.00204; abs.error = 5.3e-03 
#> n.iter = 1.5e+06 
#>    value = 0.00209; abs.error = 3.6e-03 
#> n.iter = 3.1e+06 
#>    value = 0.000823; abs.error = 2.5e-03 
#> n.iter = 6.3e+06 
#>    value = 0.000819; abs.error = 1.8e-03 
#> n.iter = 1.27e+07 
#>    value = 0.000397; abs.error = 1.2e-03 
#> n.iter = 2.27e+07 
#>    value = -5.165e-05; abs.error = 9.3e-04 

#> $value
#> [1] -5.164701e-05
#> 
#> $abs.error
#> [1] 0.0009344581
#> 
#> $cuba
#> NULL
#> 

# integrate expression
integral(parse(text = 'x'), bounds = list(x = c(0,1)))

#> n.iter = 1e+05 
#>    value = 0.4985; abs.error = 9.1e-04 

#> $value
#> [1] 0.4985093
#> 
#> $abs.error
#> [1] 0.0009124811
#> 
#> $cuba
#> NULL
#> 

# integrate function
integral(function(x) exp(x), bounds = list(x = c(0,1)))

#> n.iter = 1e+05 
#>    value = 1.720; abs.error = 1.6e-03 

#> $value
#> [1] 1.719914
#> 
#> $abs.error
#> [1] 0.001559007
#> 
#> $cuba
#> NULL
#> 

# multivariate integral
integral(function(x,y) x*y, bounds = list(x = c(0,1), y = c(0,1)))

#> n.iter = 1e+05 
#>    value = 0.2503; abs.error = 7e-04 

#> $value
#> [1] 0.250329
#> 
#> $abs.error
#> [1] 0.0006977349
#> 
#> $cuba
#> NULL
#> 

# surface of a sphere
integral('1',
         bounds = list(r = 1, theta = c(0,pi), phi = c(0,2*pi)),
         coordinates = 'spherical')

#> n.iter = 1e+05 
#>    value = 12.58; abs.error = 1.9e-02 

#> $value
#> [1] 12.5756
#> 
#> $abs.error
#> [1] 0.01915273
#> 
#> $cuba
#> NULL
#> 

# volume of a sphere
integral('1',
         bounds = list(r = c(0,1), theta = c(0,pi), phi = c(0,2*pi)),
         coordinates = 'spherical')

#> n.iter = 1e+05 
#>    value = 4.18; abs.error = 1.5e-02 

#> $value
#> [1] 4.184785
#> 
#> $abs.error
#> [1] 0.01462215
#> 
#> $cuba
#> NULL
#> 

if (FALSE) {
# efficient integration in C (requires the cubature package to be installed)
integral('1',
        bounds = list(r = c(0,1), theta = c(0,pi), phi = c(0,2*pi)),
        coordinates = 'spherical',
        method = 'cuhre',
        relTol = 1e-06,
        absTol = 1e-12)
}

##################################
# Electric charge contained in a region of space
# (see divergence theorem and Maxwell's equations)
# 

# electric potential of unitary point charge 
V <- '1/(4*pi*r)'

# electric field
E <- -1 %prod% gradient(V, c('r', 'theta', 'phi'), coordinates = 'spherical')

# electric charge
integral(E[1],
         bounds = list(r = 1, theta = c(0,pi), phi = c(0,2*pi)),
         coordinates = 'spherical')

#> n.iter = 1e+05 
#>    value = 0.999; abs.error = 1.5e-03 

#> $value
#> [1] 0.9989752
#> 
#> $abs.error
#> [1] 0.001528082
#> 
#> $cuba
#> NULL
#>